requirements.txt
django
djangorestframework
django-cors-headers
requests
django-channels
pillow
daphne
urllib3
google-auth